shifted LocalMinima() to labels Fixed index issue in RollingTransPriceMean and RollingComPriceSpreadMean Completed development and testing for RollingTransPriceMeanDiff and RollingComPriceSpreadMeanDiff
